adding cyclotmic automorphisms toy to misc
I should really figure out how to do git in the terminal, uploading through the website seems to replace everything, even the stuff that hasn't changed. 

these changes add the new page containing the javascript toy for playing with automorphisms in cyclotomic fields of order 2^n, and a page with some meta.